Move trimLongString into its own file
